[Platinum V] Title: 오아시스 재결합, Time: 72 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Platinum/3015. 오아시스 재결합/오아시스 재결합.cc
@@ -1,17 +1,18 @@
 #include <bits/stdc++.h>
 using namespace std;
+
 typedef long long ll;
 ll n, temp, ret;
 stack<pair<ll, ll>> stk;
-
-int main(){
+int main()
+{
     ios_base::sync_with_stdio(0);
     cin.tie(NULL);
     cout.tie(NULL);
     cin >> n;
     for(int i = 0; i < n; i++){
         cin >> temp;
-        int cnt = 1;
+        ll cnt = 1;
         while(stk.size() && stk.top().first <= temp){
             ret += stk.top().second;
             if(stk.top().first == temp){
